# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Products E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  comparePrice?: number | null;
  category: string;
  imageUrl: string;
  images: string[];
  inStock: boolean;
  featured: boolean;
  isUpsell: boolean;
  upsellDiscount?: number | null;
  reviewCount: number;
  rating: number;
  colors: string[];
  createdAt: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  received: "received",
  processed: "processed",
  shipped: "shipped",
} as const;

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  price: number;
  quantity: number;
  color?: string | null;
  total: number;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  shippingAddress: string;
  city: string;
  province: string;
  postalCode: string;
  country: string;
  items: OrderItem[];
  subtotal: number;
  shippingCost: number;
  tax: number;
  total: number;
  status: OrderStatus;
  isFreeOrder: boolean;
  discount: number;
  notes?: string | null;
  createdAt: string;
}

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
  color?: string;
};

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  city: string;
  province: string;
  postalCode: string;
  country: string;
  items: CreateOrderBodyItemsItem[];
  notes?: string;
  freeProductToken?: string | null;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  received: "received",
  processed: "processed",
  shipped: "shipped",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface CheckFreeProductBody {
  email: string;
}

export interface FreeProductEligibility {
  eligible: boolean;
  reason?: string | null;
}

export interface FreeProductLink {
  id: number;
  token: string;
  productId: number;
  product?: Product;
  usedByEmail?: string | null;
  usedAt?: string | null;
  createdAt: string;
}

export interface CreateFreeProductLinkBody {
  productId: number;
}

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  comparePrice?: number | null;
  category: string;
  imageUrl: string;
  images?: string[];
  inStock?: boolean;
  featured?: boolean;
  isUpsell?: boolean;
  upsellDiscount?: number | null;
  colors?: string[];
}

export type AdminStatsOrdersByStatus = {
  received: number;
  processed: number;
  shipped: number;
};

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  totalProducts: number;
  recentOrders: Order[];
  ordersByStatus: AdminStatsOrdersByStatus;
}

export type ListProductsParams = {
  category?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type ListProducts200 = {
  products: Product[];
  total: number;
};

export type GetFeaturedProducts200 = {
  products: Product[];
};

export type GetUpsellProductsParams = {
  productId?: number;
};

export type GetUpsellProducts200 = {
  products: Product[];
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  limit?: number;
  offset?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  received: "received",
  processed: "processed",
  shipped: "shipped",
} as const;

export type ListOrders200 = {
  orders: Order[];
  total: number;
};

export type DeleteProduct200 = {
  success: boolean;
};