# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Islamic Products E-Commerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product management
  - name: cart
    description: Cart operations
  - name: orders
    description: Order management
  - name: admin
    description: Admin operations
  - name: promo
    description: Promotional logic

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                  total:
                    type: integer
                required:
                  - products
                  - total

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/bestseller products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                required:
                  - products

  /products/upsell:
    get:
      operationId: getUpsellProducts
      tags: [products]
      summary: Get upsell product recommendations
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Upsell products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                required:
                  - products

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [received, processed, shipped]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                  total:
                    type: integer
                required:
                  - orders
                  - total
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /promo/check-free-product:
    post:
      operationId: checkFreeProduct
      tags: [promo]
      summary: Check if email is eligible for free product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckFreeProductBody"
      responses:
        "200":
          description: Eligibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreeProductEligibility"

  /promo/free-product-link:
    post:
      operationId: createFreeProductLink
      tags: [promo]
      summary: Generate a free product link (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFreeProductLinkBody"
      responses:
        "201":
          description: Free product link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreeProductLink"

  /promo/free-product-link/{token}:
    get:
      operationId: getFreeProductLink
      tags: [promo]
      summary: Validate and get a free product link
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Valid link with product info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreeProductLink"
        "404":
          description: Invalid link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /admin/products:
    post:
      operationId: createProduct
      tags: [admin, products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /admin/products/{id}:
    put:
      operationId: updateProduct
      tags: [admin, products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [admin, products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
          nullable: true
        category:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        isUpsell:
          type: boolean
        upsellDiscount:
          type: number
          nullable: true
        reviewCount:
          type: integer
        rating:
          type: number
        colors:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - category
        - imageUrl
        - images
        - inStock
        - featured
        - isUpsell
        - reviewCount
        - rating
        - colors
        - createdAt

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        country:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        tax:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [received, processed, shipped]
        isFreeOrder:
          type: boolean
        discount:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerEmail
        - shippingAddress
        - city
        - province
        - postalCode
        - country
        - items
        - subtotal
        - shippingCost
        - tax
        - total
        - status
        - isFreeOrder
        - discount
        - createdAt

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        price:
          type: number
        quantity:
          type: integer
        color:
          type: string
          nullable: true
        total:
          type: number
      required:
        - id
        - productId
        - productName
        - productImage
        - price
        - quantity
        - total

    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        country:
          type: string
          default: Canada
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              color:
                type: string
            required:
              - productId
              - quantity
        notes:
          type: string
        freeProductToken:
          type: string
          nullable: true
      required:
        - customerName
        - customerEmail
        - customerPhone
        - shippingAddress
        - city
        - province
        - postalCode
        - country
        - items

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [received, processed, shipped]
      required:
        - status

    CheckFreeProductBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email

    FreeProductEligibility:
      type: object
      properties:
        eligible:
          type: boolean
        reason:
          type: string
          nullable: true
      required:
        - eligible

    FreeProductLink:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        usedByEmail:
          type: string
          nullable: true
        usedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - token
        - productId
        - createdAt

    CreateFreeProductLinkBody:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
          nullable: true
        category:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        inStock:
          type: boolean
          default: true
        featured:
          type: boolean
          default: false
        isUpsell:
          type: boolean
          default: false
        upsellDiscount:
          type: number
          nullable: true
        colors:
          type: array
          items:
            type: string
      required:
        - name
        - description
        - price
        - category
        - imageUrl

    AdminStats:
      type: object
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        pendingOrders:
          type: integer
        totalProducts:
          type: integer
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        ordersByStatus:
          type: object
          properties:
            received:
              type: integer
            processed:
              type: integer
            shipped:
              type: integer
          required:
            - received
            - processed
            - shipped
      required:
        - totalOrders
        - totalRevenue
        - pendingOrders
        - totalProducts
        - recentOrders
        - ordersByStatus