# lib/api-zod/src/generated/types/checkFreeProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Products E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckFreeProductBody {
  email: string;
}