# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Products E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyItemsItem } from "./createOrderBodyItemsItem";

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  city: string;
  province: string;
  postalCode: string;
  country: string;
  items: CreateOrderBodyItemsItem[];
  notes?: string;
  freeProductToken?: string | null;
}