# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Products E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  comparePrice?: number | null;
  category: string;
  imageUrl: string;
  images?: string[];
  inStock?: boolean;
  featured?: boolean;
  isUpsell?: boolean;
  upsellDiscount?: number | null;
  colors?: string[];
}