# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Products E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  comparePrice?: number | null;
  category: string;
  imageUrl: string;
  images: string[];
  inStock: boolean;
  featured: boolean;
  isUpsell: boolean;
  upsellDiscount?: number | null;
  reviewCount: number;
  rating: number;
  colors: string[];
  createdAt: string;
}